Guard the divisions and bit shifts that could take the server down
Division by zero. CMapList marks all 256 maps supported in its
constructor and configures the first six with a sector size of 0, so
"map is supported" never implied "map has a sector size". Every helper
that divides by it could therefore raise SIGFPE for a map the .ini never
configured: GetSectorCols/GetSectorRows, CGRect::GetSector,
CPointBase::GetSector, CSectorBase::GetBasePoint and
CSector::GetLocalTime, which also divided by GetX() without checking it.
The size is now read once, checked, and the caller falls back to sector 0
or skips the offset. CSectorBase::GetBasePoint was only protected by an
ASSERT, which is compiled out on Windows release builds.

Packet::skip clamped the negative direction but not the positive one, so
a skip past the end left m_position beyond m_length and
getRemainingLength() returned a near-SIZE_MAX value - which is what the
parse loops use to decide how much to read. Clamp both ends.

CFileObjContainer::GetObjectAt used > instead of >= against size(), so an
index of exactly size() reached vector::at() and threw out_of_range.

1 << i is undefined for i >= 31 on a 32-bit int. CResourceLink's trigger
bitmask reaches 31, and CGrayInstall::OpenFiles reaches 32 because
VERFILE_QTY is 33 - which also meant VERFILE_HUES could never be selected
through a DWORD mask. The trigger mask uses 1UL and the file mask is now
64-bit, so every VERFILE_TYPE is addressable.

+++ src/common/CGrayInst.cpp
@@ -177,7 +177,7 @@ bool CGrayInstall::OpenFile( VERFILE_TYPE i )
 	return OpenFile(m_File[i], pszTitle, OF_READ|OF_SHARE_DENY_WRITE);
 }
 
-VERFILE_TYPE CGrayInstall::OpenFiles( DWORD dwMask )
+VERFILE_TYPE CGrayInstall::OpenFiles( unsigned long long dwMask )
 {
 	ADDTOCALLSTACK("CGrayInstall::OpenFiles");
 	// Now open all the required files.
@@ -186,7 +186,7 @@ VERFILE_TYPE CGrayInstall::OpenFiles( DWORD dwMask )
 
 	for ( i = 0; i < VERFILE_QTY; i++ )
 	{
-		if ( ! ( dwMask & ( 1 << i )) ) continue;
+		if ( ! ( dwMask & ( 1ULL << i )) ) continue;
 		if ( GetBaseFileName(static_cast<VERFILE_TYPE>(i)) == NULL ) continue;
 
 		bool bFileLoaded = true;
